Fix bug in apfloat handling (times() method) axexlck 15-Oct-19 12:42 AM

+++ symja_android_library/matheclipse-core/src/main/java/org/matheclipse/core/convert/AST2Expr.java
@@ -274,7 +274,7 @@ public class AST2Expr {
 		Initializer.init();
 	}
 
-	private int fPrecision;
+	private long fPrecision;
 
 	private boolean fLowercaseEnabled;
